// Web/src/app/views/restaurant/Models/Enums/EnumTypeVM.ts
export enum EnumTypeVM {
    Shifts = 1002,
    PermissionTypes = 1000,
    SaleOrderTypes = 1004,
    MenuCategory = 1014,
    ItemCategory = 1023,
    ItemSubCategory = 1015,
    Group = 1013,
    Brand = 1007,
    Size = 1010,
    Store = 1011,
    Branch = 1009,
    Color = 1006,
    ItemType = 1012,
    UOM = 1008,
    Floor = 1016,
    Table = 1017,
    BackOffice = 1018,
    Product = 1019,
    sale = 1020,
    ItemGroup = 1021,
    MenuItemGroup = 1022,
    KDS = 1024,
    PaymentType = 1025,
    Printers = 1029,
    ReportCategories = 1030,
    Devices = 1031,
    OrderTypes = 1032,
    SubItems = 1033,
    DiscountCamapignTypes = 1034,
    Accounts = 1035,
    ModifierTypes = 1036,
    Branches = 1044,
    StaffTypes = 1047,
    PurchaseTypes = 1048,
    WeekDays = 1050,
    AccountLevels,
    ChartofAccount,
    Level5 = 1057,
    SvcChrgsPaymentTypes = 1060,
    SvcChrgsOrderTypes = 1061,
    RefundTypes = 1063,
    Country = 1070,
    City,
    ClientCategories,
    Roles,
    ReportsSubCategories,
    Attributes,
    AttributeValues,
    POStatuses = 1078,
    RNTypes,
    RNCategories,
    RequesterTypes,
    RNReasons,
    RNTypes_Internal,
    RNTypes_Sales,
    RNTypes_Purchase
}