// Web/src/app/views/security/models/user-vm.ts
export class UserVM {
    clientId:number=0
    moduleId:number=0
    id?: string;
    userName?: string;
    name?: string;
    isActive: boolean;
    passwordHash?: string;
    userPassword?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    employeeId?: number = 0;
    phoneNumber?: string;
    userPhone?: string;
    directSupervisorId?: string;
    role?: string;
    fatherName?: string;
    cnic?: string
    designation?: string
    msCardNo?: string;
    address?: string;
    discountLimit?: number
    branchId?: number = 0
    branch?: string
    roleId:number=0
}